requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.0
pydantic>=2.0.0
pyyaml>=6.0
sqlglot>=23.0.0
